modify prompt provider workflow, add agent
