+++ r2r/pipelines/core/prompt.py
@@ -1,9 +1,7 @@
 from typing import Any, Optional
+from r2r.core.providers.prompt import PromptProvider
 
-from r2r.core import DefaultPromptProvider
-
-
-class BasicPromptProvider(DefaultPromptProvider):
+class BasicPromptProvider(PromptProvider):
     BASIC_SYSTEM_PROMPT = "You are a helpful assistant."
     BASIC_TASK_PROMPT = """
     ## Task:
@@ -26,11 +24,25 @@ class BasicPromptProvider(DefaultPromptProvider):
         system_prompt: Optional[str] = None,
         task_prompt: Optional[str] = None,
     ):
-        super().__init__()
+        self.prompts: dict[str, str] = {}
         self.add_prompt(
             "system_prompt", system_prompt or self.BASIC_SYSTEM_PROMPT
         )
         self.add_prompt("task_prompt", task_prompt or self.BASIC_TASK_PROMPT)
+
+    def add_prompt(self, prompt_name: str, prompt: str) -> None:
+        self.prompts[prompt_name] = prompt
+
+    def get_prompt(
+        self, prompt_name: str, inputs: Optional[dict[str, Any]] = None
+    ) -> str:
+        prompt = self.prompts.get(prompt_name)
+        if prompt is None:
+            raise ValueError(f"Prompt '{prompt_name}' not found.")
+        return prompt.format(**(inputs or {}))
+
+    def get_all_prompts(self) -> dict[str, str]:
+        return self.prompts.copy()
 
     def get_system_prompt(self, inputs: dict[str, Any]) -> str:
         return self.get_prompt("system_prompt", inputs)
